// src/types/student.ts
export interface Student {
  id: string;
  name: string;
  grade: string;
  age: number | null;
  notes: string | null;
  image: string | null;
  parentId: string;
  parentName: string;
  schoolId: string;
  schoolName: string;
  schoolBusId: string;
  schoolBusLabel: string;
  createdAt: string;
  updatedAt: string;
}